change average point type from int  to double in entities

+++ src/Reservation.Domain/Entities/Categories/Category.cs
@@ -6,7 +6,7 @@ public class Category : BaseClass
     public string Title { get; set; }
     public string Description { get; set; }
     public string CoverImagePath { get; set; }
-    public int AveragePoint { get; set; }
+    public double AveragePoint { get; set; }
 
     // Sub Categories Rel
     public ICollection<SubCategory> SubCategories { get; set; }
